ToDo: use approach of ba869407 to handle an initial transform (IT) for an mIT:
which is needed e.g. in the case of CZIto3D @ noCTRs where the sizes of mI and fI generally do not equal
the mITs created with midas use the geometric center of the fI (last registered img)
but the application of the mIT is on the mI.
For this to work, the mI needs to be padded to be as large as the fI with the mI centered
This can either be done with IM like:

convert slides/285_s01.png -gravity center -extent `convert slides/ord/reg/000789_285_s00.tif -format "%wx%h" info:` -grayscale Average slides/ord/000790_285_s01.png

but would need its own MF exception rule

cleaner is applying an IT that transforms the original image to that position
this IT (e.g. 000790_285_s01.it.txt) can be easily created:

(Transform "AffineTransform" )
(NumberOfParameters 6 )
(CenterOfRotationPoint 0 0 )
(TransformParameters 1 0 0 1 -4663.5 -7024.5 )

-4663.5 = 1823-6486.5 # w(mI)/2 - w(fI)/2
-7024.5 = 2179-9203.5 # h(mI)/2 - h(fI)/2

and made use of in mIT (000790_285_s01.txt) with adding:

(InitialTransformParametersFileName "000790_285_s01.it.txt")